In ikev2_trans, isat_transid should be a u_int16_t, not a u_int8_t. This caused interop problems on ppc64. This is bug: https://bugzilla.redhat.com/show_bug.cgi?id=438826

+++ include/packet.h
@@ -706,7 +706,7 @@ struct ikev2_trans
 	u_int16_t isat_length;	    /* Payload length */
 	u_int8_t  isat_type;        /* transform type */
 	u_int8_t  isat_res2;
-	u_int8_t  isat_transid;     /* ID */
+	u_int16_t  isat_transid;     /* ID */
 };
 extern struct_desc ikev2_trans_desc;
